+++ tsup.config.ts
@@ -10,7 +10,7 @@
 import { defineConfig } from "tsup";
 
 export default defineConfig({
-  entry: ["src/index.ts"],
+  entry: { index: "src/index.ts", image: "src/image/index.ts" },
   format: ["esm", "cjs"],
   dts: true,
   clean: true,